chore: move path summarization from common module to job attachments (#1038)
