+++ src/deadline/job_attachments/api/__init__.py
@@ -8,7 +8,7 @@ __all__ = [
     "PathSummary",
 ]
 
-from ...common.path_utils import (
+from .._path_summarization import (
     human_readable_file_size,
     summarize_paths_by_nested_directory,
     summarize_paths_by_sequence,
